Create extra data source for account view conntroler
